requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0

# Optional dependencies for future features
# sqlalchemy==2.0.23
# redis==5.0.1
# openai==1.3.0

# Development dependencies
# pytest==7.4.3
# black==23.11.0
# flake8==6.1.0
